feat(httpclient): implement the use of query defined timeout
