+++ internal/plataform/httpclient/client.go
@@ -1,15 +1,15 @@
 package httpclient
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"github.com/b2wdigital/restQL-golang/internal/domain"
 	"github.com/b2wdigital/restQL-golang/internal/plataform/logger"
 	"github.com/pkg/errors"
 	"github.com/valyala/fasthttp"
 	"time"
 )
+
+var errExecuteRequestTimeout = errors.New("request timed out")
 
 type HttpClient struct {
 	client *fasthttp.Client
@@ -34,95 +34,35 @@ func (hc HttpClient) Do(ctx context.Context, request domain.HttpRequest) (domain
 		fasthttp.ReleaseResponse(res)
 	}()
 
-	uri := fasthttp.URI{DisablePathNormalizing: true}
-	uri.SetScheme(request.Schema)
-	uri.SetHost(request.Uri)
-	uri.SetQueryStringBytes(makeQueryArgs(request))
+	setupRequest(request, req)
 
-	uriStr := uri.String()
-	hc.log.Debug("request uri build", "uri", uriStr)
-	req.SetRequestURI(uriStr)
-
-	for key, value := range request.Headers {
-		req.Header.Set(key, value)
-	}
-
-	err := hc.client.Do(req, res)
-	if err != nil {
+	err := hc.executeWithContext(ctx, req, res)
+	switch {
+	case err == errExecuteRequestTimeout:
+		hc.log.Debug("request execution did not complete on time", "request", request)
+		return domain.HttpResponse{}, errors.Wrap(err, "request execution failed")
+	case err != nil:
 		return domain.HttpResponse{}, errors.Wrap(err, "request execution failed")
 	}
 
-	responseBody, err := hc.unmarshalBody(res)
+	response, err := makeResponse(res)
 	if err != nil {
 		return domain.HttpResponse{}, err
-	}
-
-	headers := readHeaders(res)
-
-	response := domain.HttpResponse{
-		StatusCode: res.StatusCode(),
-		Body:       responseBody,
-		Headers:    headers,
 	}
 
 	return response, nil
 }
 
-func readHeaders(res *fasthttp.Response) domain.Headers {
-	h := make(domain.Headers)
-	res.Header.VisitAll(func(key, value []byte) {
-		h[string(key)] = string(value)
-	})
+func (hc HttpClient) executeWithContext(ctx context.Context, req *fasthttp.Request, res *fasthttp.Response) error {
+	errCh := make(chan error)
+	go func() {
+		errCh <- hc.client.Do(req, res)
+	}()
 
-	return h
-}
-
-func (hc HttpClient) unmarshalBody(res *fasthttp.Response) (interface{}, error) {
-	var responseBody interface{}
-	err := json.Unmarshal(res.Body(), &responseBody)
-	if err != nil {
-		hc.log.Debug("failed to unmarshal response", "error", err, "response", string(res.Body()))
-		return nil, errors.Wrap(err, "response body decode failed")
+	select {
+	case e := <-errCh:
+		return e
+	case <-ctx.Done():
+		return errExecuteRequestTimeout
 	}
-	return responseBody, nil
-}
-
-var (
-	ampersand = []byte("&")
-	equal     = []byte("=")
-)
-
-func makeQueryArgs(request domain.HttpRequest) []byte {
-	var buf bytes.Buffer
-	for key, value := range request.Query {
-		switch value := value.(type) {
-		case string:
-			appendStringParam(buf, key, value)
-		case []interface{}:
-			appendListParam(buf, key, value)
-		}
-	}
-
-	return buf.Bytes()
-}
-
-func appendListParam(buf bytes.Buffer, key string, value []interface{}) {
-	for _, v := range value {
-		s, ok := v.(string)
-		if !ok {
-			continue
-		}
-
-		buf.Write(ampersand)
-		buf.WriteString(key)
-		buf.Write(equal)
-		buf.WriteString(s)
-	}
-}
-
-func appendStringParam(buf bytes.Buffer, key string, value string) {
-	buf.Write(ampersand)
-	buf.WriteString(key)
-	buf.Write(equal)
-	buf.WriteString(value)
 }
